refactor: Client 패키지 이동
