+++ src/main/java/com/mobble/mobbleserver/infrastructure/oauth/provider/KakaoClient.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.infrastructure.oauth;
+package com.mobble.mobbleserver.infrastructure.oauth.provider;
 
-import com.mobble.mobbleserver.account.auth.oauth.dto.response.NaverUserInfoResponse;
+import com.mobble.mobbleserver.account.auth.oauth.dto.response.KakaoUserInfoResponse;
 import com.mobble.mobbleserver.application.account.command.SocialUserInfo;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
@@ -15,14 +15,14 @@ import java.util.Map;
 
 @Component
 @RequiredArgsConstructor
-public class NaverClient {
+public class KakaoClient {
 
-    private final RestClient naverRestClient;
+    private final RestClient kakaoRestClient;
 
     public SocialUserInfo fetchUserInfo(String accessToken) {
         try {
-            Map<String, Object> response = naverRestClient.get()
-                    .uri("/v1/nid/me")
+            Map<String, Object> response = kakaoRestClient.get()
+                    .uri("/v2/user/me")
                     .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
                     .retrieve()
                     .onStatus(HttpStatusCode::isError, (req, res) -> {
@@ -31,7 +31,7 @@ public class NaverClient {
                     .body(new ParameterizedTypeReference<Map<String, Object>>() {
                     });
 
-            NaverUserInfoResponse userInfo = new NaverUserInfoResponse(response);
+            KakaoUserInfoResponse userInfo = new KakaoUserInfoResponse(response);
 
             return new SocialUserInfo(
                     userInfo.getEmail(),
